iterate on IntUiBridgeTextArea
