+++ ide-laf-bridge/src/main/kotlin/org/jetbrains/jewel/bridge/theme/IntUiBridge.kt
@@ -1,7 +1,6 @@
 package org.jetbrains.jewel.bridge.theme
 
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.CornerSize
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.ExperimentalTextApi
@@ -53,17 +52,11 @@ import org.jetbrains.jewel.ui.component.styling.LazyTreeStyle
 import org.jetbrains.jewel.ui.component.styling.SelectableLazyColumnStyle
 import org.jetbrains.jewel.ui.component.styling.SimpleListItemColors
 import org.jetbrains.jewel.ui.component.styling.SimpleListItemMetrics
-import org.jetbrains.jewel.ui.component.styling.SliderColors
-import org.jetbrains.jewel.ui.component.styling.SliderMetrics
-import org.jetbrains.jewel.ui.component.styling.SliderStyle
 import org.jetbrains.jewel.ui.component.styling.TabColors
 import org.jetbrains.jewel.ui.component.styling.TabContentAlpha
 import org.jetbrains.jewel.ui.component.styling.TabIcons
 import org.jetbrains.jewel.ui.component.styling.TabMetrics
 import org.jetbrains.jewel.ui.component.styling.TabStyle
-import org.jetbrains.jewel.ui.component.styling.TextAreaColors
-import org.jetbrains.jewel.ui.component.styling.TextAreaMetrics
-import org.jetbrains.jewel.ui.component.styling.TextAreaStyle
 import org.jetbrains.jewel.ui.component.styling.TextFieldColors
 import org.jetbrains.jewel.ui.component.styling.TextFieldMetrics
 import org.jetbrains.jewel.ui.component.styling.TextFieldStyle
@@ -206,50 +199,6 @@ private fun readGroupHeaderStyle() =
                 indent = 1.dp, // see DarculaSeparatorUI
             ),
     )
-
-private fun readTextAreaStyle(metrics: TextFieldMetrics): TextAreaStyle {
-    val normalBackground = retrieveColorOrUnspecified("TextArea.background")
-    val normalContent = retrieveColorOrUnspecified("TextArea.foreground")
-    val normalBorder = DarculaUIUtil.getOutlineColor(true, false).toComposeColor()
-    val focusedBorder = DarculaUIUtil.getOutlineColor(true, true).toComposeColor()
-    val normalCaret = retrieveColorOrUnspecified("TextArea.caretForeground")
-
-    val colors =
-        TextAreaColors(
-            background = normalBackground,
-            backgroundDisabled = Color.Unspecified,
-            backgroundFocused = normalBackground,
-            backgroundPressed = normalBackground,
-            backgroundHovered = normalBackground,
-            content = normalContent,
-            contentDisabled = retrieveColorOrUnspecified("TextArea.inactiveForeground"),
-            contentFocused = normalContent,
-            contentPressed = normalContent,
-            contentHovered = normalContent,
-            border = normalBorder,
-            borderDisabled = DarculaUIUtil.getOutlineColor(false, false).toComposeColor(),
-            borderFocused = focusedBorder,
-            borderPressed = focusedBorder,
-            borderHovered = normalBorder,
-            caret = normalCaret,
-            caretDisabled = normalCaret,
-            caretFocused = normalCaret,
-            caretPressed = normalCaret,
-            caretHovered = normalCaret,
-            placeholder = NamedColorUtil.getInactiveTextColor().toComposeColor(),
-        )
-
-    return TextAreaStyle(
-        colors = colors,
-        metrics =
-            TextAreaMetrics(
-                cornerSize = metrics.cornerSize,
-                contentPadding = PaddingValues(horizontal = 5.dp, vertical = 4.dp),
-                minSize = metrics.minSize,
-                borderWidth = metrics.borderWidth,
-            ),
-    )
-}
 
 private fun readTextFieldStyle(): TextFieldStyle {
     val normalBackground = retrieveColorOrUnspecified("TextField.background")
